[Dygraph] update YOLOv3 config (#1668)
* update YOLOv3 config
