+++ configs/yolov3_darknet.yml
@@ -12,8 +12,8 @@ use_fine_grained_loss: false
 load_static_weights: True
 
 YOLOv3:
-  anchor: AnchorYOLO
   backbone: DarkNet
+  neck: YOLOv3FPN
   yolo_head: YOLOv3Head
   post_process: BBoxPostProcess
 
@@ -21,21 +21,25 @@ DarkNet:
   depth: 53
   return_idx: [2, 3, 4]
 
+YOLOv3FPN:
+  feat_channels: [1024, 768, 384]
+
 YOLOv3Head:
-  yolo_feat:
-    name: YOLOFeat
-    feat_in_list: [1024, 768, 384]
+  anchors: [10, 13, 16, 30, 33, 23, 30, 61, 62, 45, 59, 119, 116, 90, 156, 198, 373, 326]
+  anchor_masks: [[6, 7, 8], [3, 4, 5], [0, 1, 2]]
+  loss: YOLOv3Loss
+
+YOLOv3Loss:
   ignore_thresh: 0.7
   downsample: 32
   label_smooth: true
-  anchor_per_position: 3
 
 BBoxPostProcess:
   decode:
     name: YOLOBox
     conf_thresh: 0.005
     downsample_ratio: 32
-    clip_bbox: True
+    clip_bbox: true
   nms:
     name: MultiClassNMS
     keep_top_k: 100
@@ -44,14 +48,6 @@ BBoxPostProcess:
     nms_top_k: 1000
     normalized: false
     background_label: -1
-
-
-AnchorYOLO:
-  anchor_generator:
-    name: AnchorGeneratorYOLO
-    anchors: [10, 13, 16, 30, 33, 23, 30, 61, 62, 45, 59, 119, 116, 90, 156, 198, 373, 326]
-    anchor_masks: [[6, 7, 8], [3, 4, 5], [0, 1, 2]]
-
 
 LearningRate:
   base_lr: 0.001
